+++ ts/packages/providers/cloudflare/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @composio/cloudflare
 
+## 0.1.47
+
+### Patch Changes
+
+- Updated dependencies [05ce9c6]
+  - @composio/core@0.1.47
+
+## 0.1.46
+
+### Patch Changes
+
+- Updated dependencies [ee12d25]
+- Updated dependencies [9458513]
+  - @composio/core@0.1.46
+
 ## 0.1.45
 
 ### Patch Changes